requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.14
wsproto==1.2.0
XlsxWriter==3.0.7
